Add content to the omelet web page
This version of the omelet HTML code includes the
addition of a picture of the Denver omelet dish,
a "Description" section, an "Ingredients Needed"
section, and a "Cooking Steps" section.

The picture was borrowed from the allrecipes
website with no intention to infringe any
copyright rules. It is not my picture and I am
fully aware of that.