Add more linked web pages to the home page
I created two more recipe pages with a similar
beginning structure as omelet.html. Also, I
linked these web pages to the project's home
page and organized the links as an bulleted list.

The user should be able to seamlessly access any
of the recipes and navigate back to the home page.