Add content to the coxinhas web page
This version of the coxinhas HTML code includes
the addition of a picture of the Brazilian dish
known as coxinhas, a "Description" section,
an "Ingredients Needed" section, and a "Cooking
Steps" section.

The picture was borrowed from the allrecipes
website with no intention to infringe any
copyright rules.It is not my picture and I am
fully aware of that.