Add content to the egg sandwich web page
This version of the egg sandwich HTML code
includes the addition of a picture of the classic
egg sandwich dish, a "Description" section,
an "Ingredients Needed" section, and a "Cooking
Steps" section.

The picture was borrowed from the allrecipes
website with no intention to infringe any
copyright rules.It is not my picture and I am
fully aware of that.